Moved headers and fixed include statements
