+++ observation_registration/siftgpu/include/siftgpu/CuTexImage.h
