+++ observation_registration/siftgpu/include/siftgpu/FrameBufferObject.h
