+++ observation_registration/siftgpu/include/siftgpu/GLTexImage.h
