+++ observation_registration/siftgpu/include/siftgpu/GlobalUtil.h
