+++ observation_registration/siftgpu/include/siftgpu/LiteWindow.h
