+++ observation_registration/siftgpu/include/siftgpu/ProgramGLSL.h
