+++ observation_registration/siftgpu/include/siftgpu/ProgramGPU.h
