+++ observation_registration/siftgpu/include/siftgpu/PyramidGL.h
