+++ observation_registration/siftgpu/include/siftgpu/ShaderMan.h
