+++ observation_registration/siftgpu/include/siftgpu/SiftGPU.h
