+++ observation_registration/siftgpu/include/siftgpu/SiftPyramid.h
